chore: conv bulk ingress tests to perf
